updating and refactoring unit tests issue #65
