+++ Bet.AspNetCore.sln
@@ -79,6 +79,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Bet.AspNetCore.AzureStorage
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Bet.Extensions.HealthChecks", "src\Bet.Extensions.HealthChecks\Bet.Extensions.HealthChecks.csproj", "{225B565D-F4AA-46F3-8D3D-38918AAE040E}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Bet.Extensions.UnitTest", "test\Bet.Extensions.UnitTest\Bet.Extensions.UnitTest.csproj", "{C160FA06-DA6B-4988-9CBF-C7C011F3BA27}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -177,6 +179,10 @@ Global
 		{225B565D-F4AA-46F3-8D3D-38918AAE040E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{225B565D-F4AA-46F3-8D3D-38918AAE040E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{225B565D-F4AA-46F3-8D3D-38918AAE040E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C160FA06-DA6B-4988-9CBF-C7C011F3BA27}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C160FA06-DA6B-4988-9CBF-C7C011F3BA27}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C160FA06-DA6B-4988-9CBF-C7C011F3BA27}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C160FA06-DA6B-4988-9CBF-C7C011F3BA27}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -205,6 +211,7 @@ Global
 		{C8C77D66-8B0C-4600-8013-D5A4C8FDC6DA} = {D08248C1-62F0-4011-9CB2-406FDE28EEA3}
 		{919A6ABE-1E6E-4ADB-9A00-B14514AF3E04} = {8382D62B-2BB8-4406-95FD-63AF167F5D65}
 		{225B565D-F4AA-46F3-8D3D-38918AAE040E} = {8382D62B-2BB8-4406-95FD-63AF167F5D65}
+		{C160FA06-DA6B-4988-9CBF-C7C011F3BA27} = {6064A704-979A-4156-B9A8-2ADD74BB9BAE}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {A53F67D4-7CE9-4F03-8DD4-2C00E7AE2F46}
