+++ Directory.Build.props
@@ -1,12 +1,10 @@
 <Project>
 
+  <PropertyGroup>
+    <RepoRoot>$(MSBuildThisFileDirectory)</RepoRoot>
+  </PropertyGroup>
+
   <Import Project="build\settings.props" />
   <Import Project="build\sources.props" />
-
-  <PropertyGroup Label="Ownership Settings">
-    <Company>King David Consulting LLC</Company>
-    <Authors>kdcllc</Authors>
-    <Copyright>Copyright © King David Consulting LLC</Copyright>
-  </PropertyGroup>
 
 </Project>
