--- test/Bet.AspNetCore.UnitTest/GlobalSuppressions.cs
@@ -1,1 +1,0 @@
-﻿[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:File may only contain a single type", Justification = "Tests")]
