+++ test/Bet.AspNetCore.UnitTest/Options/FakeOptions.cs
@@ -1,0 +1,25 @@
+﻿using Bet.Extensions.Options;
+
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
+
+namespace Bet.AspNetCore.UnitTest.Options
+{
+    internal class FakeOptions : IOptionsFormatter
+    {
+        public int Id { get; set; }
+
+        public string Name { get; set; }
+
+        public string Format()
+        {
+            var options = new JObject
+            {
+                { nameof(Id), Id.ToString() },
+                { nameof(Name), Name },
+            };
+
+            return options.ToString(Formatting.Indented);
+        }
+    }
+}
