+++ test/Bet.AspNetCore.UnitTest/Options/FakeOptions2.cs
@@ -1,0 +1,7 @@
+﻿namespace Bet.AspNetCore.UnitTest.Options
+{
+    internal class FakeOptions2
+    {
+        public string Title { get; set; }
+    }
+}
