+++ test/Bet.Extensions.UnitTest/DIFactory/IProcess.cs
@@ -1,4 +1,4 @@
-﻿namespace Bet.AspNetCore.UnitTest.Factory
+﻿namespace Bet.Extensions.UnitTest.DIFactory
 {
     public interface IProcess<T>
     {
